Fixing collapsable sidebar button
